Split mutation engine into default (mutable) and snapshot
