+++ packages/arbor-store/src/index.ts
@@ -1,11 +1,9 @@
 export * from "./arbor"
 export * from "./decorators"
-export * from "./engine"
+export * from "./engines"
 export * from "./errors"
 export * from "./guards"
-export * from "./handlers/array"
-export * from "./handlers/default"
-export * from "./handlers/map"
+export * from "./handlers"
 export * from "./immutable"
 export * from "./path"
 export * from "./scoping/store"
